+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 0.0.2
 
 * clarify document
+
+## 0.0.3
+
+* fix bug: increasing image size after cropping
+* add new feature: `toolbarTitle` and `toolbarColor` to customize title and color of copper `Activity`